requirements.txt
fastapi
uvicorn
python-jose
passlib
sqlalchemy
asyncpg
databases
psycopg2
pymysql
aiomysql
pillow
python-multipart
bcrypt
